requirements.txt
gcloud==0.18.3
oauth2client==4.1.3
mysql-connector==2.2.9
pandas==1.4.4
SQLAlchemy==1.4.41
PyMySQL==1.0.2
google-cloud-run==0.3.0
